+++ SOC_Particle/src/talk/recall_X.cpp
@@ -157,9 +157,9 @@ boolean recall_X(const char letter_1, BatteryMonitor *Mon, Sensors *Sen)
             if ( Sen->now>TEMP_INIT_DELAY )
             {
                 Sen->start_inj = ap.wait_inj + Sen->now;
-                Sen->stop_inj = ap.wait_inj + (Sen->now + min((unsigned long int)(ap.cycles_inj / max(sp.freq()/(2.*PI), 1e-6) *1000.), ULLONG_MAX));
+                Sen->stop_inj = ap.wait_inj + (Sen->now + min((unsigned long long)(ap.cycles_inj / max(sp.freq()/(2.*PI), 1e-6) *1000.), ULLONG_MAX));
                 Sen->end_inj = Sen->stop_inj + ap.tail_inj;
-                Serial.printf("**\n*** RUN: at %ld, %7.3f cycles %ld to %ld with %ld wait and %ld tail\n\n",
+                Serial.printf("**\n*** RUN: at %lld, %7.3f cycles %lld to %lld with %ld wait and %ld tail\n\n",
                 Sen->now, ap.cycles_inj, Sen->start_inj, Sen->stop_inj, ap.wait_inj, ap.tail_inj);
             }
             else Serial.printf("Wait%5.1fs for init\n", float(TEMP_INIT_DELAY-Sen->now)/1000.);
